Build CorePlugin Version 1.0.0.0

+++ EAS_Development_Interfaces/Models/Configuration.cs
@@ -85,6 +85,7 @@ namespace EAS_Development_Interfaces
             var name = Path.GetFileNameWithoutExtension(path);
             var _appDomain = AppDomain.CreateDomain(name);
             var runtimePath = $@"{BaseDirectory}\{name}.dll";
+            if (!File.Exists(runtimePath))
             File.Copy(path, runtimePath);
             _appDomain.Load(AssemblyName.GetAssemblyName(runtimePath));
                 return _appDomain;
